third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGeometryElement.isPointInFill-01-expected.txt

This is a testharness.js-based test.
[FAIL] SVGGeometryElement.prototype.isPointInFill, no arguments.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': 1 argument required, but only 0 present.
[FAIL] SVGGeometryElement.prototype.isPointInFill, non-finite argument.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInFill, functional test.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInFill, points on the shape boundary are inside.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInFill, 'fill-rule'.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInFill, 'visibility' and 'pointer-events' have no effect.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
[FAIL] SVGGeometryElement.prototype.isPointInFill, 'clip-rule' never overrides 'fill-rule'.
  Failed to execute 'isPointInFill' on 'SVGGeometryElement': parameter 1 is not of type 'SVGPoint'.
Harness: the test ran to completion.
